fix(openapi): validate required fields and fail closed on drift
Review follow-ups. A required scalar the upstream drops now zero-values,
so the 3F offer and auction reads and the RFQ order status reconcile
validate what they act on instead of treating "" as terminal or auction
0. The relaxer fails when the generator template stops matching, the
drift check counts an unfetchable source as drift, the issue step assigns
best effort and dedups across all comments, enumUnknownDefaultCase is
dropped so a renamed enum value stays visible, and RFQ integers map to
int64 for the 2^53-1 deadlines.

+++ internal/liquidlane/discounts/client.go
@@ -106,7 +106,7 @@ func resolvedFromSingle(s *rfqbackendinternal.ResolveDiscountResponseOneOf) *Res
 		DiscountID:        s.GetDiscountId(),
 		Discount:          termsFromModel(s.GetDiscount()),
 		SignerSignature:   s.GetSignerSignature(),
-		ProtocolDeadline:  int64(s.GetProtocolDeadline()),
+		ProtocolDeadline:  s.GetProtocolDeadline(),
 		ProtocolSignature: s.GetProtocolSignature(),
 	}
 }
@@ -117,7 +117,7 @@ func resolvedFromBatchItem(requestID string, it *rfqbackendinternal.ResolveDisco
 		DiscountID:        it.GetDiscountId(),
 		Discount:          termsFromModel(it.GetDiscount()),
 		SignerSignature:   it.GetSignerSignature(),
-		ProtocolDeadline:  int64(it.GetProtocolDeadline()),
+		ProtocolDeadline:  it.GetProtocolDeadline(),
 		ProtocolSignature: it.GetProtocolSignature(),
 	}
 }
@@ -130,7 +130,7 @@ func termsFromModel(d rfqbackendinternal.ResolveDiscountResponseOneOfDiscount) T
 		Signer:        d.GetSigner(),
 		Protocol:      d.GetProtocol(),
 		Nonce:         d.GetNonce(),
-		Deadline:      int64(d.GetDeadline()),
+		Deadline:      d.GetDeadline(),
 	}
 }
 
@@ -159,7 +159,7 @@ func (c *Client) ListDiscounts(ctx context.Context) (*List, error) {
 			CollateralDecimals: int(d.GetCollateralDecimals()),
 			Discount:           d.GetDiscount(),
 			Signer:             d.GetSigner(),
-			Deadline:           int64(d.GetDeadline()),
+			Deadline:           d.GetDeadline(),
 			MaxRate:            d.GetMaxRate(),
 			MaxAssets:          d.GetMaxAssets(),
 		})
